BUG: messageType was next checked to verify type
The messageType was not being used to verify if
a message has quick replies.  This is a quick fix
to make sure that a message is not mistaken for a
quick reply.  Another fix is needed for CMS server
so that updates to messages remove quick reply
properties when appropriate.